requirements.txt
google-adk
aiosqlite
uvicorn
